Restructure WLED integration
