+++ homeassistant/components/wled/const.py
@@ -1,7 +1,12 @@
 """Constants for the WLED integration."""
+from datetime import timedelta
+import logging
 
 # Integration domain
 DOMAIN = "wled"
+
+LOGGER = logging.getLogger(__package__)
+SCAN_INTERVAL = timedelta(seconds=5)
 
 # Attributes
 ATTR_COLOR_PRIMARY = "color_primary"
